fix: move auth configmap data generation to separate file
